restructured dashboard and added routing to profile page
