+++ client/src/pages/HomePage/index.js
@@ -5,6 +5,8 @@ import { compose } from "redux";
 import { connect } from "react-redux";
 import "./style.css";
 
+import Brand from "../../components/Brand";
+import Grid from "../../components/Grid";
 import Auth from "../../containers/Auth"
 
 class HomePage extends Component {
@@ -12,7 +14,10 @@ class HomePage extends Component {
     render() {
         return (
           <div>
-            <Auth/>
+            <Brand title='VIDI'/>
+            <Grid>
+              <Auth/>
+            </Grid>
           </div>
         )
     }
@@ -27,26 +32,3 @@ export default compose(
     reduxForm({})
 )(HomePage);
 
-// This is where you're welcomed with the sign in
-//
-// import React, { Component } from "react";
-// import SignIn from "./containers/SignIn";
-// import Brand from "./components/Brand";
-// import Wrapper from "./containers/Wrapper";
-// import Container from "./Container";
-// import "./App.css";
-
-// class App extends Component {
-//   render() {
-//     return (
-//       <Wrapper>
-//         <Brand title='VIDI' />
-//         <Container>
-//           <SignIn />
-//         </Container>
-//       </Wrapper>
-//     );
-//   }
-// }
-
-// export default App;
